changing main and adding login_screen.dart
