+++ volunteering_opportunities/lib/login_screen.dart
@@ -1,0 +1,135 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
+
+
+class Login extends StatelessWidget {
+  const Login({super.key});
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Container(
+        padding: EdgeInsets.symmetric(vertical: 30),
+        width: double.infinity,
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            colors: [
+              Color.fromARGB(255, 74, 142, 76),
+              Color.fromARGB(255, 36, 112, 175)
+            ]
+          )
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            SizedBox(height: 20,),
+            Padding(
+              padding: EdgeInsets.all(20),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  Text("Welcome", style: TextStyle(color: Colors.white, fontSize: 40),),
+                  SizedBox(height: 10,),
+                  Text("Login", style: TextStyle(color: Colors.white, fontSize: 18),)
+                ]
+              )
+            ),
+            SizedBox(height: 50,),
+            Expanded(
+              child: Container(
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  boxShadow: [BoxShadow(
+                    color: Colors.black,
+                    blurRadius: 60,
+                    offset: Offset(20, 20)
+                  )],
+                  borderRadius: BorderRadius.only(topLeft: Radius.circular(60), topRight: Radius.circular(60)),
+                ),
+                child: Padding(
+                  padding: EdgeInsets.all(30),
+                  child: Column(
+                    children: <Widget> [
+                      SizedBox(height: 40,),
+                      Container(
+                        decoration: BoxDecoration(
+                          color: Colors.white,
+                          borderRadius: BorderRadius.circular(10),
+                          boxShadow: [BoxShadow(
+                            color: Color.fromARGB(92, 19, 60, 83),
+                            blurRadius: 20,
+                            offset: Offset(0, 10)
+                          )]
+                        ),
+                        child: Column(
+                          children: <Widget>[
+                            Container(
+                              padding: EdgeInsets.all(10),
+                              decoration: BoxDecoration(
+                                border: Border(bottom: BorderSide(color: Colors.grey),),
+                              ),
+                              child: TextField(
+                                decoration: InputDecoration(
+                                    hintText: "Email",
+                                    hintStyle: TextStyle(color: Colors.grey),
+                                    border: InputBorder.none
+                                  ),
+                                ),
+                              ),
+                              Container(
+                                padding: EdgeInsets.all(10),
+                                decoration: BoxDecoration(
+                                ),
+                                child: TextField(
+                                decoration: InputDecoration(
+                                  hintText: "Password",
+                                  hintStyle: TextStyle(color: Colors.grey),
+                                  border: InputBorder.none
+                                ),
+                              ),
+                            )
+                          ],
+                        ),
+                      ),
+                      SizedBox(height: 40,),
+                      Text("Forgot Password?", style: TextStyle(color: Colors.grey),),
+                      SizedBox(height: 40,),
+                      Container(
+                        height: 50,
+                        margin: EdgeInsets.symmetric(horizontal: 100),
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(50),
+                          boxShadow: [BoxShadow(
+                            color: Color.fromARGB(100, 48, 131, 179),
+                            blurRadius: 20,
+                            offset: Offset(0, 10)
+                          )],
+                          gradient: LinearGradient(
+                          begin: Alignment.topLeft,
+                          colors: [
+                            Color.fromARGB(244, 36, 112, 175),
+                            Color.fromARGB(227, 74, 142, 76),
+                            ],
+                          ),
+                        ),
+                        child: Center(
+                          child: Text("Sign In", style: TextStyle(color: Colors.white, fontSize: 16)),
+                        ),
+                      ),
+                      SizedBox(height: 40,),
+                      Text("Or continue with Google →", style: TextStyle(color: Colors.grey),),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
